feature: add config option to allow silent check of sso token from the authority, even if no user hint is provided or available
Applications using Azure AD B2C should have single-sign-on behavior
based on the tenant's configuration.  The session is saved as a cookie
on the tenant's domain.

This adds "requiresSsoHint" to the UserAgentApplication constructor.
It has a default value of "true" in keeping with the existing behavior.
If this option is set to "false", then the acquireTokenSilent does not
reject if there is no login_hint or sid provided for the authentication
request.