fix: use authority.EndSessionEndpoint for logout
The library was using a fixed logout url which will not work in all
azure ad cases. According to the oidc specifications the log out url
is published in the discovery document as end_session_endpoint.
This information is fetched and stored in the
TenantDiscoveryResponse.EndSessionEndpoint.

As the discovery property is only required if the OpenId Provider
supports session managment. We will fall back to the original
solution when the property is undefined, null or empty.