Fix error with no application in edge
There is a problem when we try to open new window with "about:blank" in url. We get error: 
"You'll need a new app to open this about" with link to the app store with empty list of applications supports this.
I used it to wrote an Add-In to Ms Office products so popup window in edge/internet explorer is the only option to login.

What have I changed:
Move a bit lines with opening popup and navigate directly to login page, not about:blank.