Set storage cookie path to '/'
This fixes two bugs with cookie storage.

The first occurs if MSAL saves its state in a directory that is not a
parent of the directory the user is returned to. The cookie will not be
available at this path because the cookie path defaults to the current
path.

The second bug occurs if identity provider does not return from a flow
(This can occur if the user closes the tab before being redirected
back), and then a new flow is started on a different path is started.
Because the default path for cookie is the current paths, and cookies
with different paths can have the same key, the msal.state.login cookie
is written twice with the same key. Under certain conditions, the wrong
cookie can be found when returning from a flow, and the state does not
match.