Example refactor for proper concurrency support
Token requests are not properly disambiguated for target authority or other semantics, this change purposes modification to some of the code that was only considering 'scope' values to instead consider a 'request uniqueness identifier' that is resolved based on the target request and account, instead of just target scopes. It seems like previous change to support setting authority per-token request was only tested when making single token calls, so the lack of proper disambiguation was not noticed.