Feature: Allow MSAL Configuration to be set at runtime
Currently module configuration can only be set at compile time. There
are instances when configuration settings are loaded dynamically and
cannot be set after the module has already been imported in AoT mode.
Allowing the config to be a static object or a function will let the
configuration token to be set after the application has already been
compiled.