[msal-core] [CacheManager] Cache should not be retrieved if new claims are provided
In this draft PR I’m showcasing a test case that I believe should fail. I’m able to retrieve a token cached that initially did not have claims associated to it even if I send a silent flow request with a claims value that I have not used before in this scenario.