Update ProtectedResourceScopes type to allow unprotected resource for concrete request type
In the Interceptor documentation (https://github.com/AzureAD/microsoft-authentication-library-for-js/blob/dev/lib/msal-angular/docs/v2-docs/msal-interceptor.md), it is stated that it should be possible to use scope value 'null' to indicate that the resource should be unprotected, even only for the concrete request type. This Pull requests ensures that the typing follows such logic.