Fix merge group workflows (#5690)
Minor adjustments to the merge group workflows added in #5679

- Don't run beachball:check on merge group trigger. It's not currently
capable of determining which file was touched unless it was triggered by
the 'pull_request' event. It should be fine to continue to run this on
each commit of the PR as a required check before adding to the merge
queue.
- Same issue with the CI workflows, restores the separate workflow we
had for the push trigger but using merge_group instead of push.