Fixed bug in msal-node's httpClient (#5722)
We've been receiving reports of users receiving the error:

`Error: TypeError
Error Description: The "listener" argument must be of type function.
Received an instance of Object - []`

I wasn't able to reproduce the issue from any user reports, however, a
fellow Microsoft employee who works on the VSCode team received the same
error in a VSCode extension she was building. She walked me through the
build steps for her VSCode extension and reproduction steps, and then I
was finally able to reproduce the issue.

As it relates to msal-node's HttpClient: I'm not sure why this error was
being thrown, because the nodejs code where the error occurs adheres to
the documentation on NodeJS's website. Specifically, I'm talking about
[https.request](https://nodejs.org/docs/latest-v16.x/api/http.html#httprequesturl-options-callback).

The relevant code in msal-node's code base is [located
here](https://github.com/AzureAD/microsoft-authentication-library-for-js/blob/dev/lib/msal-node/src/network/HttpClient.ts#L249).
According to the docs, we can make a https request this way:
 ```https.request(url, options)```.

However, I needed to change that linked code to the following to resolve
the error:
```
const customOptions: https.RequestOptions = {
        hostname: url.hostname, // <-- new code
        path: url.pathname, // <-- new code
        method: httpMethod,
        headers: options?.headers || emptyHeaders,
    };
...
https.request(customOptions)
```

Update: I have been looking into this for a while, and the only
conclusion I can draw is that maybe the NodeJS docs are incorrect, or
this is a NodeJS bug.

---------

Co-authored-by: Robert Ginsburg <rgins16@gmail.com>