Account cache optimization (#5792)
Today on every lookup of an account in the cache we iterate through
every entry in the cache (including cache entries not belonging to
MSAL), attempt to parse it into JSON and then validate that it matches
our account shape. This is incredibly inefficient and becomes more
inefficient with each additional thing stored in browser storage.

This PR adds a new cache entry (with a static key) which contains an
array of keys to all of the accounts known to MSAL. Now when the
accounts need to be enumerated the worst case number of cache entries
touched and parsed is 1 + number of accounts.

A new configuration option is added which determines whether or not this
new cache entry should be created on initialization. This is to help
migrate apps using localStorage from versions before this change.

Follow up PR will address the same issue for id, access and refresh
tokens.

---------

Co-authored-by: Thomas Norling <thnorlin@microsoft.com>
Co-authored-by: Doğan Erişen <v-derisen@microsoft.com>