Token cache optimization (#5806)
Follow up to #5792, addressing the same problem but for token lookups.

Today on every lookup of tokens in the cache we iterate through every
entry in the cache (including cache entries not belonging to MSAL),
attempt to parse it into JSON and then validate that it matches our
credential shape. This is incredibly inefficient and becomes more
inefficient with each additional thing stored in browser storage.

This PR adds a new cache entry which contains an object of keys to all
of the tokens known to MSAL for a given clientId. Now when the tokens
need to be enumerated the worst case number of cache entries touched and
parsed is 1 + number of tokens of the given type.

Example:

```js
key: msal.token.keys.<clientId>
value: {
  idToken: ["idToken-key-1", "idToken-key-2"],
  accessToken: ["accessToken-key-1", "accessToken-key-2"],
  refreshToken: ["refreshToken-key-1"]
}
```

---------

Co-authored-by: Thomas Norling <thnorlin@microsoft.com>