Retire RedisTestApp (#5773)
The current distributed caching sample RedisTestApp is rather convoluted
with framework and wrapper details that take the attention away from
caching aspects. The replacement
[auth-code-distributed-cache](https://github.com/AzureAD/microsoft-authentication-library-for-js/pull/5772)
is leaner and structurally alike to other distributed cache samples
([on-behalf-of-distributed-cache](https://github.com/AzureAD/microsoft-authentication-library-for-js/pull/5766),
[client-credentials-distributed-cache](https://github.com/AzureAD/microsoft-authentication-library-for-js/pull/5760))
to aid understanding.

The PR removes the RedisTestApp and all references to it.