Update Issue Management Bot Actions (#6242)
The new issue bot doesn't seem to be able to create conditions based on
the current date. This means every Monday it's commenting on stale PRs
stating the next release is next week. This PR contains a few small
updates to reduce confusion and spam

- Update the stale message to no longer mention the release
- Increase the number of days to be considered stale from 7 to 14

In addition, this PR restores the issue label management (Needs
Attention and Needs Author Feedback labels)