Add storeInCache request parameter to control which tokens are persisted to the cache (#6248)
Adds new request parameter to control which tokens should be stored in
the cache. This will give us more flexibility and enable our 1P library
to use our public APIs rather than depend on and/or re-implement
internal implementation details.

Note: msal-node is out of scope for this change for now. If there's a
need to extend this feature to node in the future the param can be
removed from the "Omit" list of the request types and tests added.