Multiple msal instances support for SSO (#6217)
When apps run multiple msal instances, attempt to SSO is not made for
certain instances as there is not pre-fetched RT in the cache. Though
this can be circumvented by setting appropriate cache policy or calling
`ssoSilent` instead of `acquireTokenSilent`, certain apps cannot afford
the perf hit needed to determine the right API.
This fix will have MSAL JS attempt `sso` when RT is not found in the
cache.