fix lexical declaration in case clause, type of acquireToken, add value to the result and error refs (#5878)
Hello!
I have searched the
[issues](https://github.com/AzureAD/microsoft-authentication-library-for-js/issues)
of this repository and believe that this is not a duplicate.

Version
@azure/msal-browser
2.30.0

Environment
vue
3.2.37

`plugins/msalPlugin.ts`
String 63: There is a lexical declaration in the "case" clause. It can
be a problem because the lexical declaration is visible in the entire
switch block but it only gets initialized when it is assigned, which
will only happen if the case where it is defined is reached.
I suggest adding a block in this "case":
```typescript
case EventType.ACQUIRE_TOKEN_FAILURE: {
                    const currentAccounts = msalInstance.getAllAccounts();
                    if (!accountArraysAreEqual(currentAccounts, state.accounts)) {
                        state.accounts = currentAccounts;
                    }
                    break;
                }
```
`composition-api/useMsalAuthentication.ts`
String 6: acquireToken: Function
The Function type accepts any function-like value
I suggest explicitly define the function shape
```typescript
acquireToken: (requestOverride?: PopupRequest|RedirectRequest|SilentRequest) => Promise<void>;
```
`composition-api/useMsalAuthentication.ts`
String 68:  `if (!result && !error) {`
result and error are Refs, and to access the values we should use
result.value, error.value
I suggest: `if (!result.value && !error.value) {`

Thanks! Happy coding!

Co-authored-by: Dmytro Lukashenko <dmytro.lukashenko@levi9.com>
Co-authored-by: Thomas Norling <thomas.norling@microsoft.com>