Clear claims cache for apps that switched their cache preferences [LTS] (#6356)
Some apps fetch tokens with high frequency and changing claims (new
claim per request). Prior to #6187, these tokens over flowed the local
cache. However, there is still the issue of stale tokens to be cleared
when the `claimsBasedCachingEnabled` feature is set to false, which is
not addressed.

This PR adds a new utility function `clearTokensAndKeysWithClaims()`
which will be triggered when the application is `initialized` to clear
any old tokens cached with claims.

The initial choice is to clear the cache at token acquisition. However
apps facing this issue will not be able to fetch tokens since the
`tokenKeys` cache entry is already large and freezes the app. Hence the
choice of adding this when the app is initialized and the extra
requirement of any application on the msal v2 to specifically call
`msal.initialize()` as a pre-requisite.

Applications are mandated to call `initialize()` for `msal v3` and we
are defaulting to `claimsBasedCachingEnabled: false` in v3. Hoping both
these settings will mitigate this issue from day one in the new version.

So here is the guidance:

`msal-browser v2`: App needs to explicitly set
`claimsBasedCachingEnabled: false` and call `initialize()` to use msal
to mitigate this issue.
`msal-browser v3`: MSAL JS should handle this in the back ground.

P.S: Currently this approach is in test and once approved will push the
changes to v3.

---------

Co-authored-by: Thomas Norling <thomas.norling@microsoft.com>