Change relative import to extensioned variations for Node16 and NodeNext type support (#6380)
- chore(msal-node): add $schema props for schema validation
- test(msal-node): add support for .js extension import
- build(msal-node): remove redundant pass
- feat(msal-node): change all relative imports to .js

I've made meaningful commits, so I recommend you review them instead of
all the files at once.

The other solutions are:
1. to manually transform d.ts files after the build, which is a great
way to introduce both complexity and brittleness to the project
2. use self-imports, allowing many entry-points `import { Configuration
} from "@azure/msal-node/config/Configuration"` letting typescript and
node to each resolve it in the way they understand. But this is a
massive design-shift
3. using package.json `"imports"` entry. Just changing one problem for
another, not a real solution and doesn't scale
4. emitting 2 `d.ts` versions, `d.mts` and `d.cts`. Thats possible but
just.. redundant?

There are a few oddities, like the meme extensions importing only within
their meme domain, but `d.ts` referencing non-existent `.js` files.
Which is fine because it will just direct _Typescript_ to the correct
`d.ts` and there isn't an interface difference between cjs and esm. Also
`jest` had to be adjusted since it never got the memo that 3 LTS
versions ago nodejs started supporting and advocating for `.js`
extensions. If we ditch one of the meme extensions then this would be
visually less scary, though it should be just as safe as it is now.

closes #6377

personal note:
Please don't bike-shed this pr for months, yes it would probably break
node versions before 12 but come on. [even 16 is literally days away
from deprecation](https://github.com/nodejs/release#release-schedule),
it would be absolutely unreasonable.. This is over all a small change
with a visually large foot print