Modified refreshOn in MSAL Common Silent Flow to return cached token before refreshing in the background (#6397)
refreshOn functionality already exists in Silent Flow. That is, if a
token needs to be refreshed but is not expired, an error was thrown and
the token was refreshed. This PR contains changes that will now return
the cached token if it needs to be refreshed and is not expired, while
also refreshing the token in the background.