Set TypeScript composite: false (#6494)
The composite and project references settings in TypeScript are used,
among other things, to speed up build times when the project has already
been built previously. Unfortunately, the generated tsbuildinfo file is
either incorrect or outdated and causing subsequent builds to fail. The
only way to get ourselves out of this state is to delete the tsbuildinfo
file, which defeats the purpose of the composite/project references
settings.

This PR turns off composite and project references to address the local
build issues.