Implement preconnect to speed up /token calls, refactor static class methods to exported methods for better minification (#6550)
Adds a
[preconnect](https://developer.mozilla.org/en-US/docs/Web/HTML/Attributes/rel/preconnect)
link element to the document header in flows where we're reasonably sure
we will hit the /token endpoint (acquireTokenPopup, ssoSilent,
acquireTokenSilent after RT expiration). This tells the browser to start
resolving DNS and establishing the SSL connection so that when the
/token call is made it doesn't need to wait for these steps to complete
(can take up to 300ms). This is a short lived connection and the browser
will kill it if not used promptly so I've included a cleanup function on
a timer (10s)

Also de-classed BrowserUtils for size reduction while I was adding a new
function there anyway.