Refactor token cache entities into types (#6580)
Refactors `CredentialEntity`, `IdTokenEntity`, `AccessTokenEntity`, and
`RefreshTokenEntity` to be Types rather than a Class and moves static
class methods into functions exported to the `CacheHelpers` namespace.

Making these types and separating the class methods from the type
definition allows us to read from the cache and directly use the value
without needing to first copy each key/value pair into an instance of
the class (the `toObject` helper function). Doing it this way also
results in a small bundle size improvement.

Reviews can be focused on the `msal-common/src/cache` folder. The rest
of the changes are repetitively changing references to the affected
functions