Update Nested App Auth internal schema (#6737)
Currently Nested App Auth uses a schema that contains union types such
as response.body where the body can contain a different structure
depending on the request type. This works in JavaScript that uses a
dynamic JSON parser, but some native implementations require a strongly
typed schema. There are workarounds on native, but it is easier for all
platforms to support Nested App Auth if union types are removed from the
schema.

Also remove AccountByHomeIdRequest, AccountByLocalIdRequest, and
AccountByUsernameRequest that are not currently implemented.