Fix cachePlugin.js not resolving when cache file doesn't exist (#6713)
cachePlugin.js is used for caching in the msal-node-samples. 

If the cache file does not currently exist it crates the file, however
in that scenario it currently never resolves the promise so calling
something like acquireTokenByCode will not return on the first
invocation. Calling acquireTokenByCode again will return as in that case
the cache file has been created and the promise will be resolved

This PR fixes this by resolving the promise once the file has been
created.

Signed-off-by: Robin <hi@rablet.dev>
Co-authored-by: Hector Morales <hemoral@microsoft.com>
Co-authored-by: Thomas Norling <thomas.norling@microsoft.com>