Fix: Require is not defined in MJS (#6831)
Currently, I'm getting the below error when I use msal-node-extension
package in an ES module.

```
ReferenceError: require is not defined in ES module scope, you can use import instead
This file is being treated as an ES module because it has a '.js' file extension and 'D:\git\my-repo\node_modules\my-package\package.json' contains "type": "module". To treat it as a CommonJS script, rename it to use the '.cjs' file extension.
    at file:///D:/git/my-repo/node_modules/@azure/msal-node-extensions/dist/Dpapi.mjs:20:5
    at ModuleJob.run (node:internal/modules/esm/module_job:194:25)
 ```

To fix this issue, I'm creating the `require` when it's not defined. So that this code is safe to run in both msj and cjs

---------

Co-authored-by: Thomas Norling <thomas.norling@microsoft.com>