Reduce calls to resolveEndpointsAsync (#6838)
We are building an `Authority` object and parsing/validating endpoint
metadata at least twice in a single request - once to validate it
matches the account object passed in and again to build the final
request object. This PR refactors to do both these steps at the same
time using a single `Authority` object. Additionally, telemetry is added
to track the number of times an API is invoked over the course of a
request to identify redundancies easier.