Microsoft Entra ID rebrand (#6750)
This is the attempt to fix all the issues found in the first
[PR.](https://github.com/AzureAD/microsoft-authentication-library-for-js/pull/6747)
I have run an updated script based on the feedback I got and now those
issues present in the 1st PR I made have been fixed.

Updating all Azure AD terms to the Microsoft Entra ID terms in all md
files.

## Does this introduce a breaking change?
```
[ ] Yes
[x ] No
```

## Pull Request Type
What kind of change does this Pull Request introduce?
```
[ ] Bugfix
[ ] Feature
[ ] Code style update (formatting, local variables)
[ ] Refactoring (no functional changes, no api changes)
[x] Documentation content changes
[ ] Other... Please describe:
```

## What to Check
The updated changes to Microsoft Entra ID terminology make sense.

## Other Information
Here's a link to the Rebrand guidelines. Note that the script doesn't
have all the entries in the glossary that mine has. Let me know if you'd
like to see my version of the script.
https://review.learn.microsoft.com/en-us/entra/fundamentals/how-to-rename-azure-ad?branch=main

---------

Co-authored-by: Hector Morales <hemoral@microsoft.com>