Fix compatibility issues with Nested App Auth and msal-react (#6892)
There were a few bugs that were blocking usage of msal-react in the
Nested App Auth code path.

1. handleRedirectPromise was throwing an error, and msal-react depends
on calling this before it sets user authenticated status.
2. The function useMsalAuthentication takes request as an optional
parameter, although Nested App Auth code path was considering it a
required parameter. It was throwing exceptions if request was undefined,
and correlation id is not generated until after performance event starts
for requests.