Fix for concurrent iframe calls (#6962)
When multiple acquireTokenSilent requests are made in parallel there is
a chance that they will all need to fallback to the iframe flow. It is
both unnecessary for more than 1 iframe call to be made and results in a
perf and reliability degradation for all calls. This PR introduces a
mechanism to track in progress iframe calls and cause subsequent
requests to wait before retrying the cache and/or RT redemption.

Note: Telemetry dashboards will need to be updated after this change is
released to avoid counting awaited iframe calls against our perf metrics
more than once.

Note: This PR does not make any changes to ssoSilent - follow up work
should add ssoSilent calls to the active request tracking variable and
log warnings when more than 1 ssoSilent requests are made but should
**not** block the calls.