Node-Extensions workaround npm regression (#7022)
npm version 10.4.0 introduced a regression where it auto-adds a default
install script during publish when a binding.gyp file exists. This
install script attempts to recompile binaries upon install. This is
breaking installation of node-extensions as we ship precompiled binaries
and don't include the .gyp file in the package.

To workaround this I've added an explicit install script that simply
exits so npm doesn't add its own.