Fix `useIsAuthenticated` hook briefly returning false even though the user is authenticated (#7057)
Addresses
https://github.com/AzureAD/microsoft-authentication-library-for-js/issues/6918

The current `useIsAuthenticated` has an unnecessary `useEffect`, which
causes it to briefly return `false` in some cases even when an account
is present. See
https://react.dev/learn/you-might-not-need-an-effect#updating-state-based-on-props-or-state
and
https://react.dev/learn/you-might-not-need-an-effect#caching-expensive-calculations

This PR switches to use the `useMemo` recommendation

I added a unit test that passes now, but fails with the old
implementation:
```
 FAIL  test/hooks/useIsAuthenticated.spec.tsx
  ● withMsal tests › useAuthenticated always returns true if user has an account

    expect(jest.fn()).toHaveBeenCalledTimes(expected)

    Expected number of calls: 0
    Received number of calls: 1

      78 |         expect(await screen.findByText("Has accounts")).toBeInTheDocument();
      79 |         expect(await screen.findByText("Is authed")).toBeInTheDocument();
    > 80 |         expect(invalidAuthStateCallback).toHaveBeenCalledTimes(0);
         |                                          ^
      81 |     });
      82 | });
      83 |

      at Object.<anonymous> (test/hooks/useIsAuthenticated.spec.tsx:80:42)
```

I also had to tweak some act / await / async code in other tests,
presumably because now it takes fewer renders for the hook to start
returning the correct value. Without the tweaks a couple tests were
failing, and others printed warning about updating state outside of an
`act(...)`.

---------

Co-authored-by: Thomas Norling <thomas.norling@microsoft.com>
Co-authored-by: Jo Arroyo <joarroyo@microsoft.com>