Add additional logging for NAA initialization (#7064)
Add additional logging for NAA initialization to help diagnose if
NestAppAuthController is being used for the session, or if it is falling
back to StandardController. Also log exception that occurred in trying
to use Nested App Auth bridge to get further details on failure cases.
The change also avoids the time spent trying to initialize the bridge if
caller does not opt-in to supportsNestedAppAuth.