Client Assertion implementation now accepts an async callback as well as a string argument (#7014)
Client assertion can currently be provided by a developer as a string.
This PR allows a developer to provide an async callback (which will
resolve to a string) as the client assertion.

The client assertion initialization has been removed from
ConfidentialClient's constructor and is now initialized inside of
buildOauthClientConfiguration, called during every acquireToken call.

Note: Applied Boy-Scout-Rule on UsernamePasswordClient.spec.ts. Only the
last test in the file is relevant for this review. Now this test file is
up-to-date and mirrors ClientCredentialsClient.spec.ts and
OnBehalfOfClient.spec.ts.