Supports caching in NAA apps (#7072)
NAA apps currently always request the bridge for tokens, adding load to
the host caching hence blocking the deployment on scale. This PR adds
support for:

* Caching NAA app tokens (idToken and accessToken) in the naa app
storage (based on `cacheConfig`)
* Allows for NAA app to set `initContext` with `accountContext` to look
for specific account in cache when available.

MetaOS apps create and delete the NAA apps independent of the session.
Hence, we always assume the Bridge will add the account context and only
look for those tokens. If the account is not found, we always go for the
hub to fetch tokens.

**Minor additions:**
* AccountManager created to not duplicate account lookup code and make
them functions (saving size).
* If a token is expired while cache lookup, it is cleaned.

**Note and TBD:** 
Since there is no `logout` and naa app life cycle does not map to
session life time, cache clean up is not comprehensive. However, this PR
clears the cache when a token is identified as expired. The logout
equivalent cache clearance (if the parent logs out, nested app should
clean up cache) will be handled as a separate use case.

---------

Co-authored-by: Thomas Norling <thomas.norling@microsoft.com>