fix : msal-browser acquireTokenSilentAsync memory leak (#7154)
This pull request addresses a memory leak issue in the MSAL-Browser
library.

The issue was originally identified in Zone.js for Angular 16, where the
use of arrow functions in the visibilitychange event listeners caused
problems. Specifically, Zone.js cannot compare the arrow functions when
adding and removing the event listeners, which leads to the tasks not
being properly removed and resulting in a more important memory leaks
over time.