Fix extraQueryParameters being dropped from request (#7166)
In cases where a field included in extraQueryParameters already exists
on the query string it was being deleted from the object which prevented
it from being used on the next request i.e. removed on /authorize, no
longer exists on /token. This caused a bug in Pairwise broker and NAA
flows ultimately resulting in the server throwing an
"unauthorized_client" error.

This PR fixes this bug.

Note: Objects in JavaScript are passed by reference which is why the
original implementation caused this bug. There are other places in the
code where we similarly edit an object that has been passed in. A
separate work item has been created to address those other instances in
a separate PR & turn on a lint rule to prevent this pattern in the
future.