Deprecated SHA-1 Thumbprints and Added SHA-256 support in Auth Flow with Certificate (#7185)
SHA-1 thumbprints have been deprecated. It's now up to the developer to
pass in a SHA-256 thumbprint. A warning will be shown to the dev that
states the deprecation.

This is a quick-fix that puts the burden of providing a SHA-2 thumbprint
on the developer.

Manual testing has been completed.