Fixing window unavailable bugs for PCA and PBA (#7355)
Currently, the PCA constructor in 1P and some functions for PCA (both 1P
and 3P) and PBA throw when the window is unavailable. This creates
problems with server-side rendering. Additionally, some functions that
should throw a non_browser_environment error when the window is
unavailable (such as the acquireToken functions) throw a different
error.

This PR makes changes to prevent the PCA and PBA constructor and certain
functions from throwing solely due to the window being unavailable and
to ensure that the non_browser_environment error is thrown when
appropriate. It also adds corresponding unit tests.

---------

Co-authored-by: Hector Morales <hemoral@microsoft.com>