[v4] Remove access tokens synchronously (#7822)
Updates the remove access token cache logic to perform its steps
synchronously. This is a pre-requisite for additional work to handle
cache quota errors in the near future. This requires pushing the POP key
removal (which is asynchronous) to the background, we will track
failures around this in telemetry instead of throwing an error
developers can't resolve anyway.