Fix exception when using claims with Nested App Auth in JS Runtime environment (#7926)
JS Runtime does not support full crypto API, and is running into an
exception in hydrateCache while trying to generate a sha-256 hash of the
claims. The token response is received, but never returned to developer
because writing to cache failed. For this scenario, there is already a
cache in the host, so it is still able to avoid a network call if a
previous request with the same claim was made.