Reset Redirecturi during broker fallback (#8049)
If a user attempts a broker flow but the broker isn’t available on their
machine, the flow falls back to the browser. However, since a redirect
URI was provided for the broker flow, an error is thrown.

This PR adds a check to determine whether the broker is enabled in the
config and resets the redirect URI to an empty string instead of
throwing an error.