Add native auth e2e app (#8176)
This pull request introduces a new Native Authentication sample
application for MSAL.js, adds documentation, and updates build and
ownership configuration for the new custom authentication module. The
main changes include the addition of the sample app and its
documentation, a new build configuration for the custom authentication
UMD bundle, and updates to the code ownership file to reflect
responsibility for the new sample.

---------

Co-authored-by: Konstantin <kshabelko@microsoft.com>